Implement new scheduling logic
