+++ src/types/medication.ts
@@ -35,11 +35,9 @@ export interface DoseObject {
   name: string;
   dosage: string;
   notes: string;
-  scheduledTimes: {
-    time: string;
-    status: "pending" | "taken" | "missed";
-    canCheckIn: boolean;
-  }[];
+  canCheckIn: boolean;
+  scheduledDoseTime: string;
+  status: "pending" | "taken" | "missed";
   lastTaken: Date | null;
   repeatUnit: string;
   repeatInterval: number;
